Переделал тест на функцию подсчёта очков
